Refactor StartAssessment into navigator (#272)
