+++ src/components/Donate.tsx
@@ -1,12 +1,10 @@
+import { Linking } from 'expo';
 import React, { Component } from 'react';
 import { StyleSheet, View } from 'react-native';
-
 import { colors } from '../../theme';
 import { BrandedButton, RegularBoldText, SecondaryText } from '../components/Text';
 import i18n from '../locale/i18n';
 import { CoralBadge } from './Badge';
-
-import { Linking } from 'expo';
 import Analytics, { events } from '../core/Analytics';
 
 export default class Donate extends Component {
